Move trace models into tracing package
