+++ packages/narada-pyodide/src/narada/window.py
@@ -51,7 +51,6 @@ from narada_core.actions.models import (
     UserApprovalRequest,
     UserApprovalResponse,
     WriteGoogleSheetRequest,
-    parse_action_trace,
 )
 from narada_core.errors import (
     NaradaAgentTimeoutError_INTERNAL_DO_NOT_USE,
@@ -67,6 +66,7 @@ from narada_core.models import (
     Response,
     UserResourceCredentials,
 )
+from narada_core.tracing.model import parse_action_trace
 from pydantic import BaseModel
 from pyodide.ffi import JsProxy, create_once_callable
 from pyodide.http import pyfetch
